pyproject.toml
[tool.poetry]
name = "yellot"
version = "0.1.0"
description = ""
authors = ["ErnestoTSantos <ernesto.terra2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
black = "^24.3.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
djangorestframework-simplejwt = "^5.3.1"
channels = "^4.2.0"
channels-redis = "^4.2.1"
pytest-django = "^4.9.0"
pytest-asyncio = "^0.24.0"
factory-boy = "^3.3.1"
daphne = "^4.1.2"
drf-yasg = "^1.21.8"
websockets = "^14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
